update no error without argument

+++ trash.sh
@@ -1,5 +1,7 @@
 #! /bin/sh
 
+[  $# -eq 0 ] && exit 1
+
 [ $1 == 'trash' ] || [ $1 == 'trash/' ] && ([ -f ~/trash ] || [ -d ~/trash ]) && rm -rf ~/trash && exit 0
 
 [ ! -d ~/trash ] && mkdir ~/trash
